+++ src/lib.rs
@@ -47,6 +47,48 @@ pub enum DeletionResult {
     InvalidGeneration,
 }
 
+impl<TEntry, TGeneration> IntoIterator for GenerationalVector<TEntry, TGeneration>
+where
+    TGeneration: GenerationType,
+{
+    type Item = TEntry;
+    type IntoIter = iterators::EntryIntoIterator<TEntry, TGeneration>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        iterators::EntryIntoIterator { vec: self.data }
+    }
+}
+
+impl<'a, TEntry, TGeneration> IntoIterator for &'a GenerationalVector<TEntry, TGeneration>
+where
+    TGeneration: GenerationType,
+{
+    type Item = &'a TEntry;
+    type IntoIter = iterators::EntryIterator<'a, TEntry, TGeneration>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        iterators::EntryIterator {
+            current: 0,
+            vec: &self.data,
+        }
+    }
+}
+
+impl<'a, TEntry, TGeneration> IntoIterator for &'a mut GenerationalVector<TEntry, TGeneration>
+where
+    TGeneration: GenerationType,
+{
+    type Item = &'a mut TEntry;
+    type IntoIter = iterators::EntryMutIterator<'a, TEntry, TGeneration>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        iterators::EntryMutIterator {
+            current: 0,
+            vec: &mut self.data,
+        }
+    }
+}
+
 /// A vector whose elements are addressed by both an index and an entry
 /// generation.
 impl<TEntry, TGeneration> GenerationalVector<TEntry, TGeneration>
@@ -356,6 +398,97 @@ where
 {
     fn from(vec: Vec<TEntry>) -> Self {
         Self::new_from_vec(vec)
+    }
+}
+
+/// Iterator implementations.
+pub mod iterators {
+    use super::*;
+
+    /// Iterator for owned values.
+    pub struct EntryIntoIterator<TEntry, TGeneration>
+    where
+        TGeneration: GenerationType,
+    {
+        pub(crate) vec: Vec<GenerationalEntry<TEntry, TGeneration>>,
+    }
+
+    /// Iterator for owned values.
+    pub struct EntryIterator<'a, TEntry, TGeneration>
+    where
+        TGeneration: GenerationType,
+    {
+        pub(crate) current: usize,
+        pub(crate) vec: &'a Vec<GenerationalEntry<TEntry, TGeneration>>,
+    }
+
+    pub struct EntryMutIterator<'a, TEntry, TGeneration>
+    where
+        TGeneration: GenerationType,
+    {
+        pub(crate) current: usize,
+        pub(crate) vec: &'a mut Vec<GenerationalEntry<TEntry, TGeneration>>,
+    }
+
+    impl<TEntry, TGeneration> Iterator for EntryIntoIterator<TEntry, TGeneration>
+    where
+        TGeneration: GenerationType,
+    {
+        type Item = TEntry;
+
+        fn next(&mut self) -> Option<Self::Item> {
+            while !self.vec.is_empty() {
+                match self.vec.pop() {
+                    None => continue,
+                    Some(entry) => return entry.entry,
+                }
+            }
+
+            None
+        }
+    }
+
+    impl<'a, TEntry, TGeneration> Iterator for EntryIterator<'a, TEntry, TGeneration>
+    where
+        TGeneration: GenerationType,
+    {
+        type Item = &'a TEntry;
+
+        fn next(&mut self) -> Option<Self::Item> {
+            while self.current < self.vec.len() {
+                let entry = &self.vec[self.current];
+                self.current += 1;
+                let entry = entry.entry.as_ref();
+                if entry.is_some() {
+                    return Some(entry.unwrap());
+                }
+            }
+
+            None
+        }
+    }
+
+    impl<'a, TEntry, TGeneration> Iterator for EntryMutIterator<'a, TEntry, TGeneration>
+    where
+        TGeneration: GenerationType,
+    {
+        type Item = &'a mut TEntry;
+
+        fn next(&mut self) -> Option<Self::Item> {
+            let ptr = self.vec.as_mut_ptr();
+
+            while self.current < self.vec.len() {
+                let element = unsafe { &mut *ptr.add(self.current) };
+                let entry = element.entry.as_mut();
+                self.current += 1;
+
+                if entry.is_some() {
+                    return entry;
+                }
+            }
+
+            None
+        }
     }
 }